latest changes

+++ main.js
@@ -18,17 +18,19 @@ var portNumbers = [];
 
 // Create HTTP Proxy
 
-var proxyServer = http.createServer(function(req,res) {
-    client.rpoplpush('proxyQueue','proxyQueue', function(err,value) {
-        console.log("Switching" + value );
-        var proxy = http_proxy.createProxyServer({target: value});
-		proxy.web(req, res);
-    })
-})
+var proxy = http_proxy.createProxyServer({});
 
-proxyServer.listen(80, function() {
-	console.log("listening on port 80")
-});
+var proxyServer = http.createServer(function(req,res) {
+    client.rpoplpush('proxyQueue','proxyQueue', function(err,reply) {
+        console.log("Switching" + reply );
+        var url = 'http://localhost:' + reply
+		proxy.web(req, res, {target: url});
+    })
+}).listen(80)
+
+// proxyServer.listen(80, function() {
+// 	console.log("listening on port 80")
+// });
 
 
 // Add hook to make it easier to get all visited URLS.
@@ -132,7 +134,10 @@ app.get('/get',  function(req, res) {
 
 app.get('/spawn', function(req, res){
 
+	if(globalPortCount==0){
 	client.del("servers", function(err, reply){})
+	}
+	
 	generatePortCount = generatePortCount + 1;
 	globalPortCount = globalPortCount +1;
 	port = basePort + generatePortCount;
@@ -144,7 +149,9 @@ app.get('/spawn', function(req, res){
 
   		console.log('Example app listening at http://%s:%s', host, port)
 })
-	client.lpush("servers", parseInt(port));
+	client.lpush("servers", parseInt(port), function(err, value){
+		console.log("After spawning ", value)
+	});
 	res.send("New server created at" + port);
 })
